rename package pkg/overlay to satellite/overlay
