+++ pkg/datarepair/checker/checker.go
@@ -16,10 +16,10 @@ import (
 	"storj.io/storj/internal/sync2"
 	"storj.io/storj/pkg/datarepair/irreparable"
 	"storj.io/storj/pkg/datarepair/queue"
-	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/satellite/metainfo"
+	"storj.io/storj/satellite/overlay"
 	"storj.io/storj/storage"
 )
