+++ pkg/datarepair/checker/online_test.go
@@ -12,9 +12,9 @@ import (
 
 	"storj.io/storj/internal/testcontext"
 	"storj.io/storj/internal/testrand"
-	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/storj"
+	"storj.io/storj/satellite/overlay"
 )
 
 func TestReliabilityCache_Concurrent(t *testing.T) {
