+++ satellite/nodestats/endpoint.go
@@ -11,9 +11,9 @@ import (
 	"gopkg.in/spacemonkeygo/monkit.v2"
 
 	"storj.io/storj/pkg/identity"
-	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/satellite/accounting"
+	"storj.io/storj/satellite/overlay"
 )
 
 var (
