+++ satellite/overlay/cache.go
