+++ satellite/overlay/inspector.go
