+++ satellite/vouchers/endpoint.go
@@ -12,9 +12,9 @@ import (
 	"gopkg.in/spacemonkeygo/monkit.v2"
 
 	"storj.io/storj/pkg/identity"
-	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/signing"
+	"storj.io/storj/satellite/overlay"
 )
 
 // Config contains voucher endpoint configuration parameters
